moved Logger out to its own logger util function
