+++ src/scripts/helpers.js
@@ -1,4 +1,4 @@
-const LOG_LABEL = '[BORDER PATROL]';
+import Logger from './utils/logger.js';
 
 /**
  * Checks if the provided URL is a restricted URL.
@@ -41,32 +41,6 @@ export async function getActiveTab() {
     return {};
   }
 }
-
-/**
- * A logger that allows enabling/disabling of logs.
- *
- * @constant
- * @type {Object}
- * @property {boolean} isDebug - Enables or disables debug logging
- * @property {function} info - Logs an informational message
- * @property {function} error - Logs an error message
- * @property {function} warn - Logs a warning message
- */
-export const Logger = {
-  isDebug: false,
-  info(...args) {
-    if (this.isDebug) console.log(`%c${LOG_LABEL}`, 'color: #2374ab', ...args);
-  },
-  warn(...args) {
-    if (this.isDebug) {
-      console.warn(`%c${LOG_LABEL}`, 'color: #f1c40f', ...args);
-    }
-  },
-  error(...args) {
-    // Errors are always logged regardless of debug state
-    console.error(`%c${LOG_LABEL}`, 'color: #e74c3c', ...args);
-  },
-};
 
 /**
  * Retrieves and formats the class names of an element.
